requirements.txt
sacrebleu==2.0.0
requests==2.22.0
urllib3==1.25.6
numpy==1.17.2
pyarrow==0.15.1
tqdm==4.46.0
dataclasses
boto3==1.10.18
fastprogress==0.1.21
pandas==0.25.1
botocore==1.13.18
torch==1.8.1
fire==0.2.1
beautifulsoup4==4.10.0
imageio==2.9.0
ipython==7.28.0
Pillow==8.3.2
rarfile==4.0
backoff==1.10.0
iso-639==0.4.5
sentencepiece

pyicu
transformers==4.5.1
tokenizers==0.10.3
